Break strategies up in to files
